# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AddCommentBody": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "AddCommentBody",
        "type": "object"
      },
      "AddCommentResult": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "$ref": "#/components/schemas/CaseComment"
          }
        },
        "required": [
          "comment"
        ],
        "title": "AddCommentResult",
        "type": "object"
      },
...
      "AnalysisResult": {
        "additionalProperties": false,
        "description": "Complete analysis output for an uploaded dataset.",
        "properties": {
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "anomalies": {
            "items": {
              "$ref": "#/components/schemas/AnomalyFlag"
            },
            "title": "Anomalies",
            "type": "array"
          },
          "behavioral_insights": {
            "items": {
              "$ref": "#/components/schemas/BehavioralInsight"
            },
            "title": "Behavioral Insights",
            "type": "array"
          },
          "benford_chi_squared": {
            "minimum": 0.0,
            "title": "Benford Chi Squared",
            "type": "number"
          },
          "benford_digits": {
            "items": {
              "$ref": "#/components/schemas/BenfordDigitResult"
            },
            "title": "Benford Digits",
            "type": "array"
          },
          "benford_is_suspicious": {
            "default": false,
            "title": "Benford Is Suspicious",
            "type": "boolean"
          },
          "benford_p_value": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Benford P Value",
            "type": "number"
          },
          "completed_at": {
            "format": "date-time",
            "title": "Completed At",
            "type": "string"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "graph_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphAnalysisResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "investigation_leads": {
            "items": {
              "$ref": "#/components/schemas/InvestigationLead"
            },
            "title": "Investigation Leads",
            "type": "array"
          },
          "outlier_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Outlier Count",
            "type": "integer"
          },
          "outlier_pct": {
            "default": 0.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Outlier Pct",
            "type": "number"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "risk_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "type": "integer"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "total_anomalies": {
            "minimum": 0.0,
            "title": "Total Anomalies",
            "type": "integer"
          },
          "total_transactions": {
            "minimum": 0.0,
            "title": "Total Transactions",
            "type": "integer"
          },
          "velocity_spikes": {
            "items": {
              "$ref": "#/components/schemas/VelocitySpike"
            },
            "title": "Velocity Spikes",
            "type": "array"
          }
        },
        "required": [
          "analysis_id",
          "dataset_id",
          "completed_at",
          "total_transactions",
          "total_anomalies",
          "risk_score",
          "risk_level",
          "benford_chi_squared",
          "benford_p_value"
        ],
        "title": "AnalysisResult",
        "type": "object"
      },
...
      "CaseComment": {
        "additionalProperties": false,
        "description": "A timestamped note attached to a fraud case by an analyst.",
        "properties": {
          "author_id": {
            "title": "Author Id",
            "type": "string"
          },
          "author_name": {
            "title": "Author Name",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "comment_id": {
            "title": "Comment Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          }
        },
        "required": [
          "comment_id",
          "case_id",
          "author_id",
          "author_name",
          "body",
          "created_at"
        ],
        "title": "CaseComment",
        "type": "object"
      },
      "CaseDatasetDetail": {
        "additionalProperties": false,
        "properties": {
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "row_count": {
            "minimum": 0.0,
            "title": "Row Count",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/DatasetStatus"
          },
          "uploaded_at": {
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "dataset_id",
          "name",
          "uploaded_at",
          "row_count",
          "status"
        ],
        "title": "CaseDatasetDetail",
        "type": "object"
      },
      "CaseDisposition": {
        "enum": [
          "confirmed-fraud",
          "false-positive",
          "inconclusive",
          "referred-to-law-enforcement"
        ],
        "title": "CaseDisposition",
        "type": "string"
      },
      "CasePriority": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "CasePriority",
        "type": "string"
      },
      "CaseStatus": {
        "enum": [
          "open",
          "investigating",
          "escalated",
          "resolved",
          "closed"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
      "CreateCaseCommand": {
        "additionalProperties": false,
        "properties": {
          "assigned_analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Id"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CasePriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_score": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "source_type": {
            "$ref": "#/components/schemas/WorkflowSourceType",
            "default": "scenario"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "summary"
        ],
        "title": "CreateCaseCommand",
        "type": "object"
      },
...
      "CreateCaseResult": {
        "additionalProperties": false,
        "properties": {
          "case": {
            "$ref": "#/components/schemas/FraudCase"
          }
        },
        "required": [
          "case"
        ],
        "title": "CreateCaseResult",
        "type": "object"
      },
...
      "DatasetStatus": {
        "enum": [
          "uploaded",
          "analyzing",
          "completed",
          "failed"
        ],
        "title": "DatasetStatus",
        "type": "string"
      },
...
      "FraudCase": {
        "additionalProperties": false,
        "description": "A persistent fraud investigation case with lifecycle management.",
        "properties": {
          "alert_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Alert Count",
            "type": "integer"
          },
          "assigned_analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Id"
          },
          "assigned_analyst_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Name"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "comment_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Comment Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "disposition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseDisposition"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/CasePriority",
            "default": "medium"
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "risk_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "type": "integer"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "sla_deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Deadline"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/WorkflowSourceType",
            "default": "scenario"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus",
            "default": "open"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "source_id",
          "title",
          "risk_score",
          "risk_level",
          "summary",
          "created_at",
          "updated_at"
        ],
        "title": "FraudCase",
        "type": "object"
      },
...
      "GetCaseResult": {
        "additionalProperties": false,
        "properties": {
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "case": {
            "$ref": "#/components/schemas/FraudCase"
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/CaseComment"
            },
            "title": "Comments",
            "type": "array"
          },
          "dataset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseDatasetDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "dataset_transactions": {
            "items": {
              "$ref": "#/components/schemas/UploadedTransaction"
            },
            "title": "Dataset Transactions",
            "type": "array"
          },
          "investigation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvestigationCase"
              },
              {
                "type": "null"
              }
            ]
          },
          "investigator_notes": {
            "items": {
              "$ref": "#/components/schemas/InvestigatorNote"
            },
            "title": "Investigator Notes",
            "type": "array"
          },
          "related_alerts": {
            "items": {
              "$ref": "#/components/schemas/FraudAlert"
            },
            "title": "Related Alerts",
            "type": "array"
          },
          "scenario_transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionRecord"
            },
            "title": "Scenario Transactions",
            "type": "array"
          }
        },
        "required": [
          "case"
        ],
        "title": "GetCaseResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvestigatorNote": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "note_id": {
            "title": "Note Id",
            "type": "string"
          },
          "subject_customer_id": {
            "title": "Subject Customer Id",
            "type": "string"
          }
        },
        "required": [
          "note_id",
          "subject_customer_id",
          "author",
          "created_at",
          "body"
        ],
        "title": "InvestigatorNote",
        "type": "object"
      },
...
      "ListCasesResult": {
        "additionalProperties": false,
        "properties": {
          "cases": {
            "items": {
              "$ref": "#/components/schemas/FraudCase"
            },
            "title": "Cases",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total_count": {
            "minimum": 0.0,
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "cases",
          "total_count",
          "page",
          "page_size"
        ],
        "title": "ListCasesResult",
        "type": "object"
      },
...
      "RiskLevel": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
...
      "UpdateCaseStatusBody": {
        "additionalProperties": false,
        "properties": {
          "disposition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseDisposition"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "UpdateCaseStatusBody",
        "type": "object"
      },
      "UpdateCaseStatusResult": {
        "additionalProperties": false,
        "properties": {
          "case": {
            "$ref": "#/components/schemas/FraudCase"
          }
        },
        "required": [
          "case"
        ],
        "title": "UpdateCaseStatusResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowSourceType": {
        "enum": [
          "scenario",
          "dataset"
        ],
        "title": "WorkflowSourceType",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "A fraud triage workspace for uploaded transaction datasets, with persistent alerts and cases, plus reference scenario investigations for validation and rule calibration. Statistical and behavioral analysis remain the source of truth, while the optional copilot layer explains the results in operator-facing language.",
    "title": "Relational Fraud Intelligence",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cases": {
      "get": {
        "description": "Returns a paginated list of fraud cases with optional status, priority, and analyst filters.",
        "operationId": "list_cases_api_v1_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CasePriority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCasesResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List fraud cases",
        "tags": [
          "Cases"
        ]
      },
      "post": {
        "description": "Creates a new fraud case linked to either a reference scenario or an uploaded dataset.",
        "operationId": "create_case_api_v1_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCaseResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a fraud case",
        "tags": [
          "Cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "operationId": "get_case_api_v1_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCaseResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get case details",
        "tags": [
          "Cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/comments": {
      "post": {
        "description": "Appends an analyst note or observation to a fraud case.",
        "operationId": "add_case_comment_api_v1_cases__case_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCommentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddCommentResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add a comment to a case",
        "tags": [
          "Cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/status": {
      "patch": {
        "description": "Transitions a case through its lifecycle: open, investigating, escalated, resolved, or closed.",
        "operationId": "update_case_status_api_v1_cases__case_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCaseStatusBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCaseStatusResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update case status",
        "tags": [
          "Cases"
        ]
      }
    },
...
  },
...
}